% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELRSepTests-package.R
\docType{package}
\name{ELRSepTests-package}
\alias{ELRSepTests-package}
\title{ELRSepTests: Empirical Likelihood Ratio Tests for Degrees of Separability in the Covariance Structure of Two-Way Data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Alexander Petersen \email{petersen@stat.byu.edu}

Authors:
\itemize{
  \item Garritt Page
  \item Andrew Pope
}

}
\keyword{internal}
