% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELRSepTests.R
\name{ELRSepTests}
\alias{ELRSepTests}
\title{Conduct empirical likelihood ratio tests for a sample of two-way data}
\usage{
ELRSepTests(
  X,
  tt1 = NULL,
  tt2 = NULL,
  JTest = 2L,
  LTest = 2L,
  nullHyp = c("ParSep", "WkSep", "Sep"),
  B = 500L,
  mnBoot = dim(X)[1],
  LSmax = 100L,
  ELctrl = melt::el_control(maxit = 10000L, maxit_l = 1000L),
  FVEthres = 0.99,
  thin = TRUE
)
}
\arguments{
\item{X}{n-by-M1-by-M2 array of data values.  The first direction indexes the
n observations, with \code{X[i,,]} consisting of a M1-by-M2 discretization of the
two-way data for the i-th observational unit}

\item{tt1}{optional observation grid if the second direction of X indexes
to functional data of length M1 (default is NULL)}

\item{tt2}{optional observation grid if the second direction of X indexes
to functional data of length M2 (default is NULL)}

\item{JTest}{integer or nondecreasing integer vector in 1:M1 indicating the first
direction eigenfunctions to test, must be the same length as LTest.
Default is 2L}

\item{LTest}{integer or nondecreasing integer vector in 1:M2 indicating the second
direction eigenfunctions to test, must be the same length as JTest.
Default is 2L}

\item{nullHyp}{subset of c('ParSep', 'WkSep', 'Sep') indicating which null
hypotheses to test (default is all)}

\item{B}{number of bootstrap samples for computing P values (default is 500)}

\item{mnBoot}{integer no larger than n giving the size of bootstrap samples
to be drawn - an m-out-of-n bootstrap.  Default is \code{mnBoot = n}
for a full bootstrap}

\item{LSmax}{for finding the MELE under separability, the maximum number of line
search steps to take (default = 100L)}

\item{ELctrl}{an object of class ControlEL defining options for EL estimation;
see el_control (default is el_control(maxit = 10000L, maxit_l = 100L))}

\item{FVEthres}{threshold to determine number of basis functions to use when drawing
bootstrap samples that are similar to the observed data (default is 0.99)}

\item{thin}{logical indicator for thinning output.  If thin = FALSE, then all
output below is returned; if thin = TRUE, the default, then tStatsBoot is set to NA.}
}
\value{
list with the following elements
- tStats: matrix of empirical likelihood test statistic values, with
rows corresponding to different null hypotheses and columns
to distinct pairs \verb{(J, L)}
- ELoptInfo: List of EL optimization information for each type of separability
that corresponds to a null or alternative according to nullHyp.
See optim slot from fitted ELT object for details.
Each element itself is a list corresponding to information
for fitting a given type of separability with secondary index
corresponding to distinct pairs \verb{(J, L)}.
- bootPval: matrix of bootstrap p-values, with different null hypotheses
indexing the rows and distinct pairs \verb{(J, L)} indexing columns
- tStatsBoot: list of bootstrap test statistics, \code{tStatsBoot[[a]]} is a matrix
of bootstrap test statistics for the corresponding null
hypothesis with rows indexed by bootstrap samples and
columns indexed by distinct pairs \verb{(J, L)}. Set to NA if
thin = TRUE
}
\description{
Empirical likleihood ratio tests for nested separability structures using null-transformed bootstrap calibration
}
