% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMBExp.R
\name{getMBExp}
\alias{getMBExp}
\title{Compute components of marginal tensor basis expansion of two-way data}
\usage{
getMBExp(
  X,
  tt1 = 1:dim(X)[[2]],
  tt2 = 1:dim(X)[[3]],
  J = length(tt1),
  L = length(tt2),
  useFVE = FALSE,
  FVEthres = 0.99
)
}
\arguments{
\item{X}{n-by-M1-by-M2 array of data values.  The first direction indexes the
n observations, with \code{X[i,,]} consisting of a M1-by-M2 discretization of the
two-way data for the i-th observational unit}

\item{tt1}{optional observation grid if the second direction of X indexes
to functional data of length M1 (default is 1:M1)}

\item{tt2}{optional observation grid if the second direction of X indexes
to functional data of length M2 (default is 1:M2)}

\item{J}{number of eigenvalues to extract in the first direction. Will attempt
to find J eigenvalues, but will only retain those that are positive.
Default is M1, but this will almost always be too many.}

\item{L}{number of eigenvalues to extract in the second direction. Will attempt
to find L eigenvalues, but will only retain those that are positive.
Default is M2, but this will almost always be too many.}

\item{useFVE}{logical indicating whether or not to use cumulative FVEs to determine
J and L.  If true, the values of \code{JFVE} and \code{LFVE} that achieve at
least 1 - (1 - FVEthres)/2 are computed, and the number of eigenvalues
extracted are \code{Jext = min(J, JFVE)} and \code{Lext = min(L, LFVE)} (default is FALSE)}

\item{FVEthres}{threshold for overall FVE required (default is 0.99)}
}
\value{
list with five elements
Psi - M1-by-J matrix of basis estimates for the first partial trace
operator (second direction is traced out), where J is the
number of positive eigenvalues retained
Phi - M2-by-L matrix of basis estimates for the second partial trace
operator (first direction is traced out), where L is the
number of positive eigenvalues retained
scrs - n-by-JL array of scores with columns indexed in mirror
dictionary order (1, 1), (2, 1),...(1, L),...(J, L).  That is,
if column m is associated with pair (j, l), then \code{scrs[i, m]}
is the score of (centred) \code{X[i,,]} in the direction
of the tensor product of \code{Psi[j,]} and \code{Phi[,l]}
Xbar - M1-by-M2 cross-sectional mean of X across first index
Lambda - JL-by-JL covariance matrix with row and column indices
corresponding to the column ordering of scrs
cumFVE - List of length two containing cumulative FVEs for first and
second directions.  The length of each list is equal to the
number of positive eignevalues of the corresponding partial
trace operator and may be greater than J or L
}
\description{
Uses functionality of covsep package to find the marginal bases, then expands
the data covariance in its tensor basis
}
