% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrapzVec.R
\name{getTrapzVec}
\alias{getTrapzVec}
\title{Compute trapezoidal integraion weights for grid vector tt}
\usage{
getTrapzVec(tt)
}
\arguments{
\item{tt}{observation grid for the functional data}
}
\value{
vector of same length as tt giving weights for trapezoidal integration,
}
\description{
If f is a vector representing discrete functional observations at the points
in tt, then sum(w*f) is the trapezoidal integral approximation
}
