% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SepELOpt.R
\name{SepELOpt}
\alias{SepELOpt}
\title{Outer-layer optimization for empirical likelihood estimates under separability constraint}
\usage{
SepELOpt(
  scrsAug,
  JTest,
  LTest,
  gammaStrt,
  betaStrt,
  aStrt,
  mOtr = 200L,
  mInr = 25L,
  tolOtr = 1e-06,
  tolInr = 1e-06,
  LSmax = 100L,
  verb = FALSE
)
}
\arguments{
\item{scrsAug}{augmented marginal basis scores for computing estimating equations}

\item{JTest}{number of first direction basis elements being tested}

\item{LTest}{number of second direction basis elements being tested}

\item{gammaStrt}{JTest starting values for gamma, simplicial vector of normalized first direction eigenvalues}

\item{betaStrt}{LTest starting values for beta, simplicial vector of normalized second direction eigenvalues}

\item{aStrt}{starting values for a, the trace of Lambda}

\item{mOtr}{maximum number of outer iterations (default = 200L)}

\item{mInr}{maximum number of inner iterations (default = 25L)}

\item{tolOtr}{convergence tolerance for outer layer (default = 1e-06)}

\item{tolInr}{convergence tolerance for inner layer (default = 1e-06)}

\item{LSmax}{maximal number of step size reductions before terminating line search (default = 100L)}

\item{verb}{logical indicator for printing status (default = FALSE)}
}
\value{
Object of class EL from package 'melt'
}
\description{
Implements outer-layer optimization (over space of separabe parameters)
}
