% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaldSepTests.R
\name{WaldSepTests}
\alias{WaldSepTests}
\title{Conduct Wald-type separability tests for a sample of two-way data}
\usage{
WaldSepTests(
  X,
  tt1 = 1:dim(X)[[2]],
  tt2 = 1:dim(X)[[3]],
  JTest = 2L,
  LTest = 2L,
  nullHyp = c("ParSep", "WkSep", "Sep"),
  B = 500L
)
}
\arguments{
\item{X}{n-by-M1-by-M2 array of data values.  The first direction indexes the
n observations, with \code{X[i,,]} consisting of a M1-by-M2 discretization of the
two-way data for the i-th observational unit}

\item{tt1}{optional observation grid if the second direction of X indexes
to functional data of length M1 (default is 1:M1)}

\item{tt2}{optional observation grid if the second direction of X indexes
to functional data of length M2 (default is 1:M2)}

\item{JTest}{number of projections to test in the first direction, cannot be more
than M1 (default is 2)}

\item{LTest}{number of projections to test in the second direction, cannot be more
than M2 (default is 2)}

\item{nullHyp}{subset of c('ParSep', 'WkSep', 'Sep') indicating which null
hypotheses to test (default is all)}

\item{B}{number of bootstrap samples for computing P values (default is 500)}
}
\value{
list with the following elements
- tStats: test statistic value(s) - squared length of (studentized)
vector of Wald statistics corresponding to specified null
(note that these will BE NA for testing separability since
covsep doesn't return them)
- bootRes:
* tStats: list of length B vector of bootstrap test statistics, one vector per test
(note that these will be NA for testing separability since covsep doesn't return them)
* bootPval: bootstrap p-values, one per test
}
\description{
Wald tests of Aston et al. (full studentization) and Lynch and Chen (no studentization)
for assessing separability structures using bootstrap calibration
}
