% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawBoot.R
\name{drawBoot}
\alias{drawBoot}
\title{Draw a null-transformed bootstrap sample}
\usage{
drawBoot(MBExp, nullType)
}
\arguments{
\item{MBExp}{list output of getMBExp on original data}

\item{nullType}{one of 'ParSep', 'WkSep', 'Sep' indicating the null hypothesis
under which to draw the bootstrap sample}
}
\value{
3D array of null-bootstrapped two-way data with dimensions n-by-p-by-s
}
\description{
Modifes the marginal basis expansion to match a particular separability structure and appropriate bootstrap sample
}
