% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELRSepTests.R
\name{ELRSepTests}
\alias{ELRSepTests}
\title{Conduct empirical likelihood ratio tests for a sample of two-way data}
\usage{
ELRSepTests(
  X,
  tt1 = 1:dim(X)[[2]],
  tt2 = 1:dim(X)[[3]],
  JTest = 2L,
  LTest = 2L,
  nullHyp = c("ParSep", "WkSep", "Sep"),
  B = 500L,
  LSmax = 100L,
  ELctrl = melt::el_control(maxit = 10000L, maxit_l = 1000L),
  FVEthres = 0.99
)
}
\arguments{
\item{X}{n-by-M1-by-M2 array of data values.  The first direction indexes the
n observations, with \code{X[i,,]} consisting of a M1-by-M2 discretization of the
two-way data for the i-th observational unit}

\item{tt1}{optional observation grid if the second direction of X indexes
to functional data of length M1 (default is 1:M1)}

\item{tt2}{optional observation grid if the second direction of X indexes
to functional data of length M2 (default is 1:M2)}

\item{JTest}{number of projections to test in the first direction, cannot be more
than M1 (default is 2)}

\item{LTest}{number of projections to test in the second direction, cannot be more
than M2 (default is 2)}

\item{nullHyp}{subset of c('ParSep', 'WkSep', 'Sep') indicating which null
hypotheses to test (default is all)}

\item{B}{number of bootstrap samples for computing P values (default is 500)}

\item{LSmax}{for finding the MELE under separability, the maximum number of line
search steps to take (default = 100L)}

\item{ELctrl}{an object of class ControlEL defining options for EL estimation;
see el_control (default is el_control(maxit = 10000L, maxit_l = 100L))}

\item{FVEthres}{threshold to determine number of basis functions to use when drawing
bootstrap samples that are similar to the observed data (default is 0.99)}
}
\value{
list with the following elements
- tStats: test statistic value(s) - minus twice the constrained empirical log-likelihood ratio
corresponding to specified null and alternative
- loglr - log empirical likelihood ratio value for each type of separability
- ELoptInfo: EL optimization information for each test (see optim slot from fitted ELT object).
For testing weak separability and separability, there is information for both the
constrained null fit and specific alternative
- bootRes: bootstrap calibration info list (if cal = 'bootstrap', otherwise is NA)
* tStats: list of length B vector of bootstrap test statistics, one vector per test
* loglr: list of length B, each giving the log empirical likelihood ratio for each type of separability
* ELoptInfo: list of lists EL optimization information for each bootstrap, same structure as
ELoptInfo above for each element
* bootPval: bootstrap p-values, one per test
}
\description{
Empirical likleihood ratio tests for nested separability structures using null-transformed bootstrap calibration
}
