% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getELTestData.R
\name{getELTestData}
\alias{getELTestData}
\title{Compute and arrange products of marginal tensor basis scores for use in ELR tests}
\usage{
getELTestData(scrs, J, L, JTest = J, LTest = L)
}
\arguments{
\item{scrs}{n-by-JL matrix of scores in marginal tensor product basis}

\item{J}{number of first direction basis elements used in the expansion}

\item{L}{number of second direction basis elements used in the expansion}

\item{JTest}{number of first direction basis elements to include in test data}

\item{LTest}{number of second direction basis elements to include in test data}
}
\value{
matrix of augmented data with columns organized as follows:
cross-products for different direction 2 bases - LTest(LTest - 1)JTest^2/2 columns for ParSep
cross-products for different direction 1 bases but direction 2 basis - JTest(JTest - 1)LTest/2 columns for WkSep
squares - LTest*JTest columns for Sep
}
\description{
Compute and arrange products of marginal tensor basis scores for use in ELR tests
}
