% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaldSepTests.R
\name{WaldSepTests}
\alias{WaldSepTests}
\title{Conduct Wald-type separability tests for a sample of two-way data}
\usage{
WaldSepTests(
  X,
  tt1 = NULL,
  tt2 = NULL,
  JTest = 2L,
  LTest = 2L,
  nullHyp = c("ParSep", "WkSep", "Sep"),
  B = 500L,
  thin = TRUE
)
}
\arguments{
\item{X}{n-by-M1-by-M2 array of data values.  The first direction indexes the
n observations, with \code{X[i,,]} consisting of a M1-by-M2 discretization of the
two-way data for the i-th observational unit}

\item{tt1}{optional observation grid if the second direction of X indexes
to functional data of length M1 (default is NULL)}

\item{tt2}{optional observation grid if the second direction of X indexes
to functional data of length M2 (default is NULL)}

\item{JTest}{integer or nondecreasing integer vector in 1:M1 indicating the first
direction eigenfunctions to test, must be the same length as LTest.
Default is 2L}

\item{LTest}{integer or nondecreasing integer vector in 1:M2 indicating the second
direction eigenfunctions to test, must be the same length as JTest.
Default is 2L}

\item{nullHyp}{subset of c('ParSep', 'WkSep', 'Sep') indicating which null
hypotheses to test (default is all)}

\item{B}{number of bootstrap samples for computing P values (default is 500)}

\item{thin}{logical indicator for thinning output.  If thin = FALSE, then all
output below is returned; if thin = TRUE, the default, then tStatsBoot is set to NA.}
}
\value{
list with the following elements
- tStats: matrix of test statistic value(s) with rows corresponding to
different null hypotheses and columns to distinct pairs \verb{(J, L)}
of test dimensions in each direction. These will be NA for
testing separability since \code{covsep} doesn't return them
- bootPval: bootstrap p-value matrix, with rows corresponding to null
hypotheses and columns to distinct pairs \verb{(J, L)}.  Will be
NA for testing separability
- tStatsBoot: list of length equal to the different null hypotheses being
tested. \code{tStatsBoot[p]} has B rows and columns corresponding
to distinct paris \verb{(J, L)} of test dimensions in each direction
}
\description{
Wald tests of Aston et al. (full studentization) and Lynch and Chen (no studentization)
for assessing separability structures using bootstrap calibration
}
